fix: corriger les anomalies arbitrées en réunion métier du 20/07
- Gamme/Série (sale.order): liste de Série filtrée par la Gamme
  sélectionnée (relation x_serie_mtn -> x_gamme_mtn).
- Commercial (Employé): champ recréé sur res.partner (existait déjà,
  custom/models/res_partner.py), propagé automatiquement sur le devis
  (sale.order.x_studio_commercial_1, calculé) et sur l'Affaire
  (nouveau champ x_studio_commercial_id sur x_affaire) -- FMA/Janneau
  n'utilise pas le commercial natif Odoo (licences), uniquement la
  notion d'Employé.
- Mode de règlement: nouveau champ x_studio_mode_de_rglement_id sur
  x_affaire, calculé depuis x_studio_mode_de_rglement_dsa du contact lié
  (distinct de la Condition de paiement standard).
- Chronologie du devis: les groupes "Chronologie"/"Informations
  principales"/"Informations supplémentaires" étaient insérés en dehors
  du groupe natif sale_header, s'empilant en pleine largeur au lieu de
  rejoindre la grille à 2 colonnes native -- xpath corrigé.
- _rec_name + translate=True sur 16 champs "x_name" à travers 8 modèles
  Studio portés (x_affaire_stage, x_gamme_mtn, x_serie_mtn,
  x_capacite_par_poste, x_delai_entre_operatio(_line), x_reglements,
  famille x_remise, x_purchase_order_line_35a7b,
  x_account_move_line_803a2) : les listes déroulantes affichaient soit
  l'identifiant technique brut ("x_affaire_stage,1"), soit un
  dictionnaire de traduction non résolu ({'en_US': ..., 'fr_FR': ...}).
  Aucune migration de données nécessaire dans les deux cas.
- Achats: Is XML Created, creation time et send to SFTP ajoutés à la
  fiche commande (existaient déjà en code, visibles en liste seulement).
- Vitrage: nouveau champ calculé x_is_glazing_order (catégorie produit
  "Remplissage") pour n'afficher remise/commentaires/dimensions que sur
  les commandes vitrage ; "Contrat cadre" (requisition_id) masqué
  partout (confirmé inutile).
- Capacité par poste: durée min/max passées d'Integer à Float (un seuil
  comme 80,1h était tronqué à 80, faussant le comparateur ">80h").
- Congés/capacité: hr.leave déclenche désormais un recalcul immédiat des
  mrp.capacity.week concernées (create/write/unlink) au lieu d'attendre
  le cron quotidien qui, de plus, ne retraitait jamais les semaines déjà
  passées -- élargi aux 30 derniers jours en filet de sécurité.

+++ fma_studio_models/models/x_capacite_par_poste.py
@@ -28,8 +28,13 @@ class XCapaciteParPoste(models.Model):
 
     x_active = fields.Boolean(string="Actif", default=True)
     x_name = fields.Char(string="Description", required=True, translate=True)
-    x_studio_dure_max = fields.Integer(string="Durée Max")
-    x_studio_dure_min = fields.Integer(string="Durée Min")
+    # Entier -> Décimal : un seuil comme "80,1h" était tronqué à 80 (type
+    # entier), faisant échouer le comparateur ">80h" utilisé dans
+    # mrp_capacity_planning/models/mrp_production.py::_get_effective_duration_hours
+    # (retour métier, réunion du 2026-07-20 : "non lié à la migration v19,
+    # même comportement en v17", correctif ponctuel demandé).
+    x_studio_dure_max = fields.Float(string="Durée Max")
+    x_studio_dure_min = fields.Float(string="Durée Min")
     x_studio_nbre_ressources = fields.Integer(string="Nbre Ressources")
     x_studio_poste = fields.Many2one("mrp.workcenter", string="Poste")
     x_studio_sequence = fields.Integer(string="Séquence")
